начала разбирать conftest: перенесла функции генерации тестовых данных в генераторы, синглтон хранилища использованных данных - в хэлперы, и начала делать фикстуру-фабрику для пользователя
