+++ tests/api_tests/test_project_type.py
@@ -6,7 +6,7 @@ import allure
 from main.api.requests.checked_crud_request import CheckedRequest
 from main.api.requests.unchecked_crud_request import UncheckedRequest
 from main.api.requests.endpoints import Endpoint
-from tests.conftest import generate_test_child_project, generate_test_project, generate_test_user, \
+from tests.factories.generators import generate_test_user, generate_test_project, generate_test_child_project, \
     generate_test_copy_project
