+++ tests/factories/user_factory.py
@@ -1,0 +1,7 @@
+import pytest
+
+
+@pytest.fixture
+def user_factory(specifications):
+    def user_builder(**kwargs):
+        pass
